move get available cities function to the filter service
